requirements.txt
customtkinter>=5.2.0
selenium>=4.15.0
undetected-chromedriver>=3.5.0
ntplib>=0.4.0
Pillow>=10.0.0
